add linc-cleanup-sockets.
2002-05-21  Michael Meeks  <michael@ximian.com>

	* src/Makefile.am: add linc-cleanup-sockets.

	* src/cleanup.c: add.

	* src/linc-connection.c
	(linc_connection_writev): make status an int,
	so we return the enum more pleasantly.

+++ linc2/src/linc-connection.c
@@ -622,7 +622,7 @@ linc_connection_writev (LINCConnection       *cnx,
 			const LINCWriteOpts  *opt_write_opts)
 {
 	QueuedWrite qw;
-	glong       status;
+	int         status;
 
 	/* FIXME: need an option to turn this off ? */
 	if (cnx->options & LINC_CONNECTION_NONBLOCKING) {
